Abstracted method to class UnixCronControl
